Bugfix - No CLDataIOs are created for scalar variables.
When getDataIO(..) is making a CLDataIO from a Type, it first tries to treat it as a tuple.  If this fails it then tries to match the type to the appropriate scalar CLDataIO subtype.  Unfortunately TupleAnalysis.getTupleInfo(..) throws an unhandled RuntimeException which bypasses the the fallback to scalar CLDataIOs.  This fix first matches the type to the known scalars and then falls back to trying a tuple.